Eliminate backup files from root folder
By default, any file created by the "newFile" command adds the file to the root directory on Google Drive.  By adding it to the configured almanac folder, it does not get removed from the root folder.

This fix will get the file handle of the new file, add it to the almanac folder, and remove the file from the root folder.

(It is important to note that files in Google Drive are tagged with folders much like messages are tagged in Gmail.  Therefore, files can be in multiple folders at a time.)