Make inclusion of chart configurable
The chart is included only if there is the following setting in the config tab:

Include Chart in Almanac | 1 | Set | Include a chart that contains the goal, total and 7-day average in the daily email

For backward compatibility, I added a method that takes a default value for the config and doesn't stop execution if it can't find the value in the spreadsheet. This value defaults to 0, so if a user copies this new version, he shouldn't see any change in behavior.